refactor : main 페이지 ui 리팩토링
