+++ src/components/common/AutoCarousel/AutoCarousel.module.scss
@@ -16,13 +16,17 @@
     align-items: center;
 
     &__image {
-      width: 7rem;
-      height: 8rem;
+      width: 8rem;
+      height: 10rem;
       object-fit: contain;
       border-radius: 10px;
       //   box-shadow: 0px 5px 15px rgba(0, 0, 0, 0.3);
 
       z-index: 1;
+      @include desktop {
+        width: 5rem;
+        height: 6rem;
+      }
     }
 
     &__nameTag {
@@ -40,6 +44,11 @@
       display: flex;
       justify-content: center;
       align-items: center;
+
+      @include desktop {
+        width: 100px;
+        height: 1.5rem;
+      }
     }
   }
 }
