+++ src/pages/main/centerview/LiveNewsItem.module.scss
@@ -6,9 +6,12 @@
   gap: 15px;
 
   width: 100%;
-  height: 100%;
+  flex: 1;
+  min-height: 0;
+
   @include desktop {
-    height: 7rem;
+    flex: 1;
+    min-height: 0;
   }
 
   &__image {
