+++ src/pages/main/centerview/styles/centerview.module.scss
@@ -57,7 +57,7 @@
       &__dataSection {
         width: 80%;
         // height: 60%;
-        height: 30rem;
+        height: 20rem;
         display: flex;
         flex-direction: row;
         justify-content: space-between;
