+++ src/pages/main/index.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useMemo, useState } from "react";
 import styles from "./styles/main.module.scss";
 import MenuView from "./menuview/index";
-import CenterView from "./centerview/index";
+import CenterView from "./centerview/CenterView";
 import PersonalView from "./personalview/index";
 import { AnimatePresence, motion } from "framer-motion";
 import { useRecoilValue, useSetRecoilState, useRecoilState } from "recoil";
